app/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-multipart
torch
requests
Pillow